allow to generate a full font glyph (although crashes for now)
